small refactoring and moved out from models calculate offset to its own module
