+++ lib/kaminari/models/mongoid_extension.rb
@@ -1,4 +1,5 @@
 require 'kaminari/models/mongoid_criteria_methods'
+require 'kaminari/models/page_extensions'
 
 module Kaminari
   module MongoidExtension
@@ -19,7 +20,7 @@ module Kaminari
       include Kaminari::ConfigurationMethods
 
       included do
-        self.send(:extend, ClassMethods)
+        self.send(:extend, Kaminari::PageExtensions)
         # Fetch the values at the specified page number
         #   Model.page(5)
         scope Kaminari.config.page_method_name, Proc.new {|num|
@@ -27,21 +28,6 @@ module Kaminari
         } do
           include Kaminari::MongoidCriteriaMethods
           include Kaminari::PageScopeMethods
-        end
-      end
-
-      module ClassMethods
-        def calculate_offset(num)
-          num = default_per_page * ([num.to_i, 1].max - 1)
-          goto_page = Kaminari.config.out_of_range
-
-          return 0 if goto_page == :first && out_of_range?(num)
-          return scoped.count - default_per_page if goto_page == :last && out_of_range?(num)
-          num
-        end
-
-        def out_of_range?(num)
-          num + 1 > scoped.count
         end
       end
